Support for inline editing of files for insert subcommand
The insert subcommand now accepts an '--inline' option that allows you
to edit a yaml file inline as opposed to dumping it to stdout.

Also added a new TempSampleCopy class that can be used to easily work
with temporary files in test cases (i.e. it can be used to test the
--inline behavior).