Fix for parsing certain yaml files
yamlpal was throwing an error when parsing certain complex yamlfiles.
The actual fix is not fully understood at this point as it has to do the
internals of how the pyYAML parser parses the yaml file, but this commit
adds a case in the parsing where we stop parsing further if we encounter
a yamlnode of a primitive type.