Initial support for working with yaml dictionaries and lists
Yamlpal now allows you to search for dictionary or list elements. In the
past you could only search for
leafs in a yaml document, i.e. you had to specify a yamlpath string that
resolved to an atomic key-value pair or
item in a list. Now you can search for e.g. "foo/bar" where "bar" can
resolve to a complex dictionary or list.

Notes:
- This still needs more work wrt how insertion behavior should work in
  these cases
- We don't have yaml pretty printing for dictionary or lists yet, for
  now we do json pretty printing because
  it was an easy workaround
- Also added PyYAML to requirements.txt