Initial support for end line numbers
Yamlpal is now able to determine the line number on which an element
ends. This is particularly useful for finding the compound elements like
dictionaries and lists which are spread across multiple lines.

This capability is required for future additions that will add better smarts
around finding/inserting/modifying/deleting content inside of compound elements.